Fixed class fields to map database fields Added default inserts for database

+++ storage_project/src/main/java/ru/stepan/ponomarev/storage_project/dto/UserInfo.java
@@ -13,7 +13,6 @@ import lombok.NoArgsConstructor;
  * Info of user accounts to be shown for user
  */
 public class UserInfo {
-
     String username;
-    float balance;
+    double balance;
 }
